update and delete routes fixed

+++ BikeStores.Api/Controllers/BrandController.cs
@@ -3,10 +3,12 @@ using BikeStores.Application.Interfaces;
 using BikeStores.Domain.Dtos;
 using BikeStores.Domain.Models;
 using Microsoft.AspNetCore.Mvc;
+using Swashbuckle.AspNetCore.Annotations;
+
 namespace BikeStores.Api.Controllers
 {
     [ApiController]
-    [Route("api/[controller]")]
+    [Route("api/brands")]
     public class BrandController : ControllerBase
     {
         private readonly IBrandRepository _brandRepository;
@@ -21,7 +23,8 @@ namespace BikeStores.Api.Controllers
         }
 
         [HttpGet]
-        public async Task<IActionResult> GetAllBrands()
+        [SwaggerResponse(200, "Success")]
+        public async Task<IActionResult> GetAllBrandsAsync()
         {
             var brands = await _brandRepository.GetAllAsync();
             _logger.LogInformation($"Retrieved {brands.Count()} brands");
@@ -29,7 +32,9 @@ namespace BikeStores.Api.Controllers
         }
 
         [HttpGet("{id}")]
-        public async Task<IActionResult> GetBrand(int id)
+        [SwaggerResponse(200, "Success")]
+        [SwaggerResponse(404, "Not found")]        
+        public async Task<IActionResult> GetBrandAsync(int id)
         {
             var brand = await _brandRepository.GetAsync(id);
             if (brand == null)
@@ -43,28 +48,55 @@ namespace BikeStores.Api.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> CreateBrand([FromBody] CreateBrandDto createBrandDto)
+        [SwaggerResponse(200, "Success")]
+        [SwaggerResponse(409, "Already existing")]
+        [SwaggerResponse(400, "Validation errors occured")]
+        public async Task<IActionResult> CreateBrandAsync([FromBody] CreateBrandDto createBrandDto)
         {
+            var existingBrand = await _brandRepository.GetByNameAsync(createBrandDto.BrandName);
+            if (existingBrand != null)
+            {
+                string loggedErrorMessage = $"Brand with name '{createBrandDto.BrandName}' already exists";
+                _logger.LogInformation(loggedErrorMessage);
+                return Conflict(loggedErrorMessage);
+            }
             var brand = _mapper.Map<Brand>(createBrandDto);
             var newBrandId = await _brandRepository.CreateAsync(brand);
             _logger.LogInformation($"Created brand with ID {newBrandId}");
 
-            return CreatedAtAction(nameof(GetBrand), new { id = newBrandId }, _mapper.Map<BrandDto>(brand));
+            return CreatedAtAction(nameof(GetBrandAsync), new { id = newBrandId }, _mapper.Map<BrandDto>(brand));
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateBrand(int id, [FromBody] UpdateBrandDto updateBrandDto)
+        [SwaggerResponse(204, "No content")]
+        [SwaggerResponse(404, "Not found")]
+        [SwaggerResponse(400, "Validation errors occured")]
+        public async Task<IActionResult> UpdateBrandAsync(int id, [FromBody] UpdateBrandDto updateBrandDto)
         {
-            var brand = _mapper.Map<Brand>(updateBrandDto);
-            await _brandRepository.UpdateAsync(id, brand);
-            _logger.LogInformation($"Updated brand with ID {id}");
+            var existingBrand = await _brandRepository.GetAsync(id);
+            if (existingBrand == null)
+            {
+                _logger.LogInformation($"Brand with ID {id} not found");
+                return NotFound();
+            }
 
+            var newBrand = _mapper.Map<Brand>(updateBrandDto);
+            await _brandRepository.UpdateAsync(id, newBrand);
+            _logger.LogInformation($"Updated brand with ID {id}");
             return NoContent();
         }
 
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteBrand(int id)
+        [SwaggerResponse(204, "No content")]
+        [SwaggerResponse(404, "Not found")]
+        public async Task<IActionResult> DeleteBrandAsync(int id)
         {
+            var existingBrand = await _brandRepository.GetAsync(id);
+            if (existingBrand == null)
+            {
+                _logger.LogInformation($"Brand with ID {id} not found");
+                return NotFound();
+            }
             await _brandRepository.DeleteAsync(id);
 
             _logger.LogInformation($"Deleted brand with ID {id}");
